+++ src/data-workspace/custom-form/replace-td-node.jsx
@@ -1,11 +1,7 @@
 import { attributesToProps } from 'html-react-parser'
 import React from 'react'
 import { IndicatorTableCell } from '../indicators-table-body/indicator-table-cell.jsx'
-import { CustomFormTotalCell } from './custom-form-total-cell.jsx'
-
-const replaceTotalCell = (dataElementId) => (
-    <CustomFormTotalCell dataElementId={dataElementId} />
-)
+import { replaceTotalCell } from './custom-form-total-cell.jsx'
 
 const replaceIndicatorCell = (indicatorId, metadata) => {
     const {